refactor: rxws messages operator is now named consume
