+++ packages/rxws/src/operators/index.js
@@ -2,5 +2,5 @@ export {default as broadcast} from './broadcast';
 export {default as client} from './client';
 export {default as conduit} from './conduit';
 export {default as connections} from './connections';
+export {default as consume} from './consume';
 export {default as disconnections} from './disconnections';
-export {default as messages} from './messages';
